adding test and change test file name
